Revert "Feat/#54 영화 검색 기능 구현"
