+++ src/main/java/com/insidemovie/backend/api/member/entity/Member.java
@@ -82,7 +82,6 @@ public class Member extends BaseTimeEntity {
     }
 
     // 회원 정지
-    @Column(nullable = false)
     public void setBanned(boolean banned) {
         this.isBanned = banned;
     }
