--- src/main/java/com/insidemovie/backend/api/movie/dto/movie/MoviePagedResponseDTO.java
@@ -1,16 +1,0 @@
-package com.insidemovie.backend.api.movie.dto.movie;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-
-public class MoviePagedResponseDTO<T> {
-    private List<T> content;
-    private int pageNumber;
-    private int pageSize;
-    private long totalElements;
-    private int totalPages;
-}
