+++ src/main/java/com/insidemovie/backend/api/movie/repository/MovieRepository.java
@@ -1,10 +1,9 @@
 package com.insidemovie.backend.api.movie.repository;
 
 import com.insidemovie.backend.api.movie.entity.Movie;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
@@ -13,10 +12,7 @@ import java.util.Optional;
 public interface MovieRepository extends JpaRepository<Movie, Long> {
 
     List<Movie> findAllByTmdbMovieIdIn(Collection<Long> tmdbIds);
+    //List<Movie> findByTmdbMovieId(Long tmdbIds);
 
     Optional<Movie> findByTmdbMovieId(Long tmdbMovieId);
-
-    Page<Movie> findByTitleContainingIgnoreCase(String title, Pageable pageable);
-
-    boolean existsByTmdbMovieId(Long tmdbMovieId);
 }
