--- src/main/resources/application-local.yml
@@ -1,19 +1,0 @@
-spring:
-  h2:
-    console:
-      enabled: true
-
-  datasource:
-    driver‑class‑name: org.h2.Driver
-    url: 'jdbc:h2:file:~/insidemovie;AUTO_SERVER=TRUE;DB_CLOSE_DELAY=-1'
-    username: sa
-    password:
-
-  jpa:
-    hibernate:
-      ddl-auto: update
-    database-platform: org.hibernate.dialect.H2Dialect
-    show-sql: true
-
-  batch:
-    initialize-schema: never
